data-pipelines/dags/kafka-app/requirements.txt
apache-airflow==2.4.0
kafka-python==2.0.2
pandas==1.3.3
psycopg2-binary==2.9.3
boto3==1.20.10
requests==2.26.0
